Implemented uploadImage and fetchImageById routes.

+++ routes/user.js
@@ -1,6 +1,7 @@
 import express from "express"
 import jwt from "jsonwebtoken"
 import fs from "fs/promises"
+import bcrypt from "bcrypt"
 import "dotenv/config"
 import User from "../models/user.js"
 
@@ -8,58 +9,83 @@ const userRouter = express.Router()
 
 userRouter.post("/register", async (req, res) => {
   try {
-    //Get secret key and user name if provided
+    //Get secret key,username and hashed password.
     const secretKey = process.env.JWT_SECRET_KEY ?? "THISISFUN"
-    const userName = req.body.userName ?? "Unknown User"
-    console.log("req.body.userName: ", req.body)
+    const { userName, password } = req.body
+    if (!userName) {
+      return res.status(400).json({ error: "Please provide the username." })
+    }
+    if (!password) {
+      return res.status(400).json({ error: "Please provide the password." })
+    }
+    const passwordToString = password.toString()
 
-    //Get current JSON DB
+    const saltRounds = 15
+    const hashedPassword = await bcrypt.hash(passwordToString, saltRounds)
+
+    //Get current JSON DB as an object.
     const allUsersJSON = await fs.readFile("db/users.json", "utf-8")
     const allUsersObject = JSON.parse(allUsersJSON)
 
-    //Create new User object
-    const newUserId = allUsersObject.users.length + 1
-    const newUserObject = new User(newUserId, userName)
+    //Check whether username already exists.
+    const existingUser = allUsersObject.users.find((user) => user.username === userName)
+    if (existingUser) {
+      return res.status(400).json({ error: "Username already exists. Choose a different username." })
+    }
 
-    //Push user into users object and write JSON DB.
+    //Create new User object.
+    const newUserId = (allUsersObject.users?.length ?? 0) + 1
+    const newUserObject = new User(newUserId, userName, hashedPassword)
+
+    //Push user into users object and write on the JSON DB.
     allUsersObject.users.push(newUserObject)
     await fs.writeFile("db/users.json", JSON.stringify(allUsersObject, null, 2))
 
     //Sign with secret key and send token as json.
     const jwtToken = jwt.sign({ userId: newUserId }, secretKey)
     console.log(`A new user has registered with ID = ${newUserId} & username = '${userName}'`)
-    res.status(200).json({ jwtToken })
+    res.status(200).json({ jwtToken, message: "Sucessfully registered!" })
   } catch (err) {
     console.log("Error while registering", err)
     res.status(500).json({ error: "Failed to register." })
   }
 })
 
-userRouter.post("/loginWithId", async (req, res) => {
+userRouter.post("/login", async (req, res) => {
   try {
-    //Get secret key and user Id
+    //Get secret key and user Id.
     const secretKey = process.env.JWT_SECRET_KEY ?? "THISISFUN"
-    const userId = req.body.userId
+    const { userName, password } = req.body
+    const passwordToString = password.toString()
 
-    if (!userId) {
-      return res.status(400).json({ error: "Please provide the userId." })
+    if (!userName) {
+      return res.status(400).json({ error: "Please provide the username." })
+    }
+    if (!password) {
+      return res.status(400).json({ error: "Please provide the password." })
     }
 
-    //Get current JSON DB
+    //Get current JSON DB.
     const allUsersJSON = await fs.readFile("db/users.json", "utf-8")
     const allUsersObject = JSON.parse(allUsersJSON)
 
-    //Check if the user exists
-    const userExists = allUsersObject.users.some((user) => user.id === userId)
-
-    if (!userExists) {
+    //Check if the user exists and get their ID.
+    const user = allUsersObject.users.find((user) => user.username === userName)
+    if (!user) {
       return res.status(400).json({ error: "Such a user does not exist. Please register first." })
     }
 
+    //Check whether password matches.
+    const passwordMatches = await bcrypt.compare(passwordToString, user.password)
+    if (!passwordMatches) {
+      return res.status(401).json({ error: "Your password is incorrect." })
+    }
+
     //Sign with secret key and send token as json.
+    const userId = user.id
     const jwtToken = jwt.sign({ userId }, secretKey)
     console.log(`ID ${userId} requested for their token!`)
-    res.status(200).json({ jwtToken })
+    res.status(200).json({ jwtToken, message: "Sucessfully logged in!" })
   } catch (err) {
     console.log("Error while logging in.", err)
     res.status(500).json({ error: "Failed to login." })
